move pessimistic prewrite to action
