+++ src/storage/mvcc/reader/reader.rs
@@ -506,7 +506,7 @@ mod tests {
     use crate::storage::kv::Modify;
     use crate::storage::mvcc::{MvccReader, MvccTxn};
 
-    use crate::storage::txn::{commit, prewrite};
+    use crate::storage::txn::{commit, pessimistic_prewrite, prewrite};
     use concurrency_manager::ConcurrencyManager;
     use engine_rocks::properties::MvccPropertiesCollectorFactory;
     use engine_rocks::raw::DB;
@@ -595,7 +595,8 @@ mod tests {
             let cm = ConcurrencyManager::new(start_ts);
             let mut txn = MvccTxn::new(snap, start_ts, true, cm);
 
-            txn.pessimistic_prewrite(
+            pessimistic_prewrite(
+                &mut txn,
                 m,
                 pk,
                 &None,
